Creates instructions for changing attributes on a dom element
Turns out the implementation of attributes in the synthetic dom
is different in deno dom and in the JSDOM from node world. I had
to not use the most useful, and probably more efficient methods
in the underlying class because they weren't implemented.

When there are intergration or e2e tests, I may swap out this
implementation for ones that are more efficient.