Maybe fixes re-render issues
Instead of the root conditionally replacing itself with the new
dom, the root gets the new dom from the parent itself. That means
it's always in sync. Since reading the dom is slow, it's going
to be a bit of a hit on performance, maybe?

Also changes the build to be a bit easier for bug releases.
package.json has the version number and the dist is always the
same name. That also means that I could run my cypress tests
easier without having to change the import statements.