Don't generate children instructions when removing or replacing node
There were a lot of instructions being generated that were maybe
messing things up, but definitely were unneccessary when a node
was removed either directly or via replace and then we were doing
those child steps too.

This skips children instructions when replacing or removing a node.
I think this can also be true for adding nodes, but haven't tried
yet.

This commit also adds a ton of tests unit tests that are maybe not
that useful, but helped sanity check for me.