Big ole fix for diff algorithm with insertions in the middle
Strategy:
* Switches to a smarter lookup algorithm over dumber going
  through the list and replacing with target
* Creates an id mapping object and infers a jsx id from:
  * key (like react/preact)
  * __source, if compiled that way
  * dom characteristics:
    - tag name
    - id
    - type
    - name
* Starts top of target and tries to move existing elements
  instead of inserting new ones from target. This should prevent
  funky focus issues. Though if you move a input child someplace
  very different in the dom, the insertAfter will cause focus to
  be lost. Going to see how often this is really an issue.
* After looking at the targets and trying to match to existing or
  insert new ones, the remaining, unhandled source elements are
  deleted.

The instructions create insertNode for move and also for inserting
something new. In the case of inserting to move, change will no-op
if it is already in that location.

Fixes a bug in the separation of events and attributes where anything
events where catching anything with 'on' anywhere in the attribute
name ... like 'dontupdate'. Now it only matches at the beginning.

Algorithm also builds all the insert/delete/move before recursing
into the node changes. This isn't real tail recursion, but it made
debugging easier.

After the instructions for a child are created, then they get sorted
so that the mutations are on top. I think this might be overly
aggressive since by the end we are sorting the whole array of changes.

Separates the node instructions compiler from the children, which
required some currying. Sorry.

Fixes a bug where target attributes were not added if there were
more target attributes than source attributes.