Event publishing for programattic navigation too:
When inside an event that, for example, completes an API call, the
developer may want to change pages in the handler. Previously, the
logic for navigating was only avaialable via elements with an href
in their ancestor chain, This changes that event name and uses the
more generic 'navigate' method for the programatic version.

Also cleans up a few tests along the way.