Adds SVG handling
One more joy moment for js standards. The creation of svg takes a
different set of node and attribute setting functions on the
document. This passes down a boolean flag from each tag down
to all the children templates, which means it assumes that devs
aren't going to add html tags into the middle of a svg. I hope
that's true.

Testing this was problematic: creating the namespaced element in the
dom implementation causes a seg fault! At first I thought it was
because of the way I was spying on the method, but I think this is
a bug in the dom implementation. I'm going to try it in my project
and see if it actually works. For now that test in skipped, and I
unit tested that the boolean is passed down properly.