Debugging svn via install on path into project
Fixes a whole bunch of type import issues created by the latest
Typescript.

Passes in the namespace correctly! That solved the seg fault
problem. The browser was able to give me a message I could debug
instead of seg fault.

There is still a bug where the only tag generated with namespace
is 'svg', no the children. So the next commit will probably look
into that problem.