Fixes bug in test/setting of svg:
Children was setting everything to the value of the parent, but
when the parent was the svg element and first in tree, it was set
by parent element to not svg. So this only does that set if it's
not already set to true.

It also adds an integration test to make sure that the svg renders
properly. It checks the height width, to figure that out.