Window/document dependency injection
I had a bit of dependency injection in the works, but alas it was
not complete. This does some window/document resolution depending on
what is passed in, and then it goes from there.

The fallback to document from the global space isn't working because
Bun is half-baked and searching the internet for how to make this
work seems impossible. So this isn't fully tested. I'd like to switch
to Vite or something else, but I don't have several years to debug
that system too (and I tried).