Changes the check for svg to lookup the tag
Turns out the parser operates from inside out, or depth first. So
the parent svg can't pass down to children the svg nature fo their
tag type. The only conflicting tag is `a` which will need to be
decorated with the svg namespace is order to be treated as svg.

Going to write some cypress tests to see how that will work with
rendering.