Lots of debugging to sort out the bug(s) without creating more
* The parent being properly passed down from tag to children was one
  possible issue. It certainly changed the bug that originally started
  this work.
* Keeping the bound dom cache up to date was the bigger issue. Inserts
  deletions and replacements at the root level were performed on the
  window dom, but not in the cached dom in the bound template. That
  template needs an up to date version for the next round of change and
  comparison. So, it's important.

The first fix attempted to add better methods to the bound class
to reconcile the differences, but it was kind of hasty and dependent
on high level integration tests. Even the unit-y tests that were
happening for the Bound template were too far away from the logic
to make it work. It led to even more bugs since I wasn't taking into
account the flattened recursive nature of the change list. Most of
change instructions aren't for the root level. So there are only a
few that we need to worry about.

This fix extracts that logic into a module and makes it more like the
perform change work, so that each edge case could be tested well. E2E
tests all worked after the unit tests were written.