Issue #6 - allow uncontrolled forms
Because we are resetting the value even when the value hasn't
changed it was seeming to interfere with creating an uncontrolled
form. I'm not sure this is a real thing, without tests or trying
to create an uncontrolled form. But the code was in and it doesn't
seem to interfere.