Bug: Issue #28 - more cloning to avoid mutating initial state (#29)
While the initial state was using `structuredClone` when setting the
initial value, it wasn't when setting the private _value attribute.
During the `update` method that took a replacement state, everything was
fine, but during a update that took an updater function mutating that
state, all hell brokes loose.

This fixes that bug. But I think we should also be making sure the the
updater form of update gets a clone of the original data. So that was
done too.