Better types (#37)
Type improvements

- Got feedback from AI about the best way to dist the types:

This adds type fields to the package.json. It also updates the build
process to use the ds vite package which is seemingly better for tree
shaking. Finally it publishs the less quirky version of types.

- Implements AI advice on how to better export types for DX

Change allows developers to import types from the library directly:
`import type { App } from 'jaxs'`. Probably not all the common
ones are currently exported, so more work needed.

The module still exports types as JaxsTypes for backwards compatibility.
Not sure what the performance implications are for this though.