BUG fix: version 0.7.3 issue with children rendering (#38)
Normalization to ensure that all the children are JXS renderable wasn't
happening high enough in the stack. That lead to a string (no render
method) showing up in these nested children instead of Text elements.

This CR moves the normalization of jsx higher up and catches this case.
It leaves the big integration-y test behind.

Another thing that I found in this bug is that some attributes were
being set to an empty string `class=""` etc. So this also adds a test
and fix to the tag/dom related code to exclude empty strings (with
trim).