Moves payload into ListenerKit:
Why? Often the event was unnecessary and so there was lots
of TS and argument nonsense managing that. Now there is a
payload attribute on the ListenerKit type, and the kit is
a template type that folks can use when they want to specify
the type.

Also fixes the cypress tests to do the right thing!