Organinizing types: moves to directory and related files
Typing for Publish function, plus a bit of clarity

I think I went back and forth about typing publish. It definitely
gets a type into it, and since users are creating that type it
makes sense to give them that type safety option. But it cascades
pretty hard. Or does it?

For example the ListenerKit has a generit T that is related to the
payload it receives when it is published. But the listener kit
provides a way to spawn other types of events with other payloads.

This pushes in some generic typing without forcing developers to
define what can't be defined.

Also combines typing for PublishToDom which has T = Event, with the
Publsih function used by the bus more generally.

- Separates out types for rerender-instructions and jxs
- Simplifies JSX collection since earlier bug fixed this type inconsistency
- Also renames to RenderableCollection
- Extracts out rendering types and renames instructions to rerendering
- Extracts out the bus related types

Most of the types sorted out. Need to retire, reconfigure updaters